requirements.txt
# Trading and data analysis
pandas>=1.5.0
numpy>=1.21.0
ta-lib>=0.4.25
ccxt>=4.0.0
pybit>=5.0.0

# API and networking
requests>=2.28.0
aiohttp>=3.8.0
websockets>=10.0

# Web scraping and data extraction
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.8.0
playwright>=1.30.0
scrapy>=2.9.0

# Database
sqlalchemy>=1.4.0

# Machine learning
scikit-learn>=1.1.0

# Utilities
python-dotenv>=0.19.0
asyncio
logging
schedule
datetime
json
time